Move code for loading key pairs into CryptoAdapter
Now, there is only one implementation of CryptoAdapter
